requirements.txt
mutagen==1.47.0
pygame-ce==2.5.3.dev2
tqdm==4.64.1
